+++ src/components/ui/BaseInput.vue
@@ -5,11 +5,24 @@
       <span v-if="label">{{ label }}</span>
     </label>
     <input
+      v-if="lazy"
+      :type="type"
+      id="search"
+      :name="name"
+      :placeholder="placeholder"
+      v-model.lazy="inputValue"
+      ref="input"
+      :disabled="disabled"
+    />
+    <input
+      v-else
       :type="type"
       id="search"
       :name="name"
       :placeholder="placeholder"
       v-model="inputValue"
+      ref="input"
+      :disabled="disabled"
     />
   </div>
 </template>
@@ -40,6 +53,35 @@ export default {
     label: {
       type: String,
       default: '',
+    },
+    reset: {
+      type: Boolean,
+      default: false,
+    },
+    lazy: {
+      type: Boolean,
+      default: false,
+    },
+    disabled: {
+      type: Boolean,
+      default: false,
+    },
+    focus: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  watch: {
+    reset(reset) {
+      if (reset) {
+        this.$refs['input'].value = '';
+      }
+    },
+    focus(focus) {
+      if (focus) {
+        this.$refs['input'].focus();
+        this.$emit('focused');
+      }
     },
   },
   computed: {