+++ tests/unit/components/home/GuessPokemon.test.js
@@ -24,7 +24,7 @@ describe('GuessPokemon', () => {
 
   it('renders correctly', () => {
     expect(wrapper.exists()).toBe(true);
-    expect(wrapper.find('.game-results').text()).toBe('Guess the Pokemon!');
+    expect(wrapper.find('#game-results').text()).toBe('Guess the Pokemon!');
     expect(wrapper.find('baseinput-stub').exists()).toBe(true);
   });
 
@@ -32,5 +32,9 @@ describe('GuessPokemon', () => {
     const playersGuess = 'Pikachu';
     wrapper.vm.setPlayersGuess(playersGuess);
     expect(wrapper.vm.playersGuess).toBe(playersGuess);
+  });
+
+  it('computes the image property correctly', () => {
+    expect(wrapper.vm.image).toBe('pikachu.png');
   });
 });