+++ CHANGELOG.md
@@ -2,7 +2,29 @@
 
 ---
 
-## [V0.1.1](https://github.com/nashaguayo/pokedex/releases/tag/V0.1.1)
+## [Unreleased](https://github.com/nashaguayo/pokedex/compare/V0.1.2...HEAD) (dd-mm-yyyy)
+
+---
+
+## [V0.1.2](https://github.com/nashaguayo/pokedex/compare/V0.1.1...V0.1.2) (10-07-2023)
+
+##### Added
+
+- Pokemon Game: limiting guesses per pokemon and showing amount of guesses in a row
+
+##### Modified
+
+- Optimized Search
+- Using store everywhere
+
+##### Fixed
+
+- Pokemon Game: input clears when getting a new pokemon
+- Pokemon Details: fixed issues with missing properties
+
+---
+
+## [V0.1.1](https://github.com/nashaguayo/pokedex/compare/V0.1.0...V0.1.1) (09-07-2023)
 
 ##### Added
 
@@ -21,7 +43,7 @@
 
 ---
 
-## [V0.1.0](https://github.com/nashaguayo/pokedex/releases/tag/V0.1.0)
+## [V0.1.0](https://github.com/nashaguayo/pokedex/releases/tag/V0.1.0) (28-06-2023)
 
 ##### Added